fix(admin): Move post types to new file and add Subscription
