+++ src/admin/components/donors/DonorsTable.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { dateI18n } from '@wordpress/date';
 import { useAdminContext, usePostsContext } from '../contexts';
 import { useEffect } from '@wordpress/element';
-import type { Campaign, Donor } from '../../../types/wp';
+import type { Donor } from '../../../types/posts';
 export const DonorsTable = (): React.ReactNode => {
 	const { setPageTitle } = useAdminContext();
 	const {
